requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
elasticsearch==8.17.1
sentence-transformers==3.4.1
pypdf==5.2.0
langchain==0.3.17